Windows: fixed various warning issues.
The API was built with everything ON except for ENABLE_EVERYTHING, ENABLE_FSL_INTEGRATORS, ENABLE_JAVA_BUILD and ENABLE_TELICEMS, which were left to OFF.

+++ sources/cellml_context/CellMLContextImplementation.cpp
@@ -569,7 +569,7 @@ void
 CDA_ModelNode::stampModifiedNow()
   throw(std::exception&)
 {
-  mTimestamp = time(0);
+  mTimestamp = (uint32_t) time(0);
 }
 
 already_AddRefd<iface::cellml_api::Model>
